pyproject.toml
[tool.poetry]
name = "todo-app"
version = "0.1.0"
description = "A simple ToDo app for your DevOps apprenticeship"
authors = []

[tool.poetry.dependencies]
python = "^3.7"
flask = "^1.1.2"
python-dotenv = "^0.14.0"
requests = "^2.27.0"
gunicorn = "^20.1.0"
pymongo = "^4.2.0"

[tool.poetry.dev-dependencies]
pytest = "^7.0.0"
mongomock = "^4.1.2"
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
